requirements.txt
Flask
numpy
Pillow
protobuf==3.20.3
tensorflow==2.15.0
gunicorn